Rudimentary inventory
Created 2 arrays, clues and found_clues
When a Clue is created it is put into the clues array and stuck to
background.

When a clue is clicked (distance not measured yet) the clue is copied
into found_clues, positon is changed, and it is unstuck from the
background.

I am still working on it.